[#159] PartyListView 코드 정리
